# frontend/src/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "LifeGrid API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/plans/{plan_id}": {
   "get": {
    "tags": [
     "redistribution"
    ],
    "summary": "Get Plan",
    "operationId": "get_plan_api_v1_plans__plan_id__get",
    "parameters": [
     {
      "name": "plan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Plan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlanOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/recommendations/{rec_id}/decision": {
   "post": {
    "tags": [
     "redistribution"
    ],
    "summary": "Decide",
    "operationId": "decide_api_v1_recommendations__rec_id__decision_post",
    "parameters": [
     {
      "name": "rec_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Rec Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DecisionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RecommendationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/transfers/{transfer_id}/pick": {
   "get": {
    "tags": [
     "redistribution"
    ],
    "summary": "Preview Pick",
    "description": "The units dispatch would pick right now (FEFO, minimum remaining life).",
    "operationId": "preview_pick_api_v1_transfers__transfer_id__pick_get",
    "parameters": [
     {
      "name": "transfer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Transfer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/UnitOut"
         },
         "title": "Response Preview Pick Api V1 Transfers  Transfer Id  Pick Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/transfers/{transfer_id}/dispatch": {
   "post": {
    "tags": [
     "redistribution"
    ],
    "summary": "Dispatch",
    "operationId": "dispatch_api_v1_transfers__transfer_id__dispatch_post",
    "parameters": [
     {
      "name": "transfer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Transfer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TransferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/transfers/{transfer_id}/receive": {
   "post": {
    "tags": [
     "redistribution"
    ],
    "summary": "Receive",
    "operationId": "receive_api_v1_transfers__transfer_id__receive_post",
    "parameters": [
     {
      "name": "transfer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Transfer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TransferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests": {
   "get": {
    "tags": [
     "requests"
    ],
    "summary": "List Requests",
    "operationId": "list_requests_api_v1_requests_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RequestOut"
         },
         "type": "array",
         "title": "Response List Requests Api V1 Requests Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "requests"
    ],
    "summary": "Create",
    "operationId": "create_api_v1_requests_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RequestIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests/{request_id}": {
   "get": {
    "tags": [
     "requests"
    ],
    "summary": "Get",
    "operationId": "get_api_v1_requests__request_id__get",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Request Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests/{request_id}/confirm": {
   "post": {
    "tags": [
     "requests"
    ],
    "summary": "Confirm",
    "operationId": "confirm_api_v1_requests__request_id__confirm_post",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Request Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests/{request_id}/cancel": {
   "post": {
    "tags": [
     "requests"
    ],
    "summary": "Cancel",
    "operationId": "cancel_api_v1_requests__request_id__cancel_post",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Request Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests/{request_id}/outcome": {
   "post": {
    "tags": [
     "requests"
    ],
    "summary": "Outcome",
    "operationId": "outcome_api_v1_requests__request_id__outcome_post",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Request Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UnitsIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/requests/{request_id}/matches": {
   "get": {
    "tags": [
     "requests"
    ],
    "summary": "Request Matches",
    "operationId": "request_matches_api_v1_requests__request_id__matches_get",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Request Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MatchOut"
         },
         "title": "Response Request Matches Api V1 Requests  Request Id  Matches Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/appeals": {
   "post": {
    "tags": [
     "requests"
    ],
    "summary": "Create Appeal",
    "operationId": "create_appeal_api_v1_appeals_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AppealIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AppealIn": {
    "properties": {
     "site_id": {
      "type": "string",
      "format": "uuid",
      "title": "Site Id"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "units": {
      "type": "integer",
      "maximum": 20.0,
      "minimum": 1.0,
      "title": "Units"
     },
     "urgency": {
      "$ref": "#/components/schemas/Urgency",
      "default": "urgent"
     },
     "hours": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 336.0,
        "minimum": 2.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Hours"
     }
    },
    "type": "object",
    "required": [
     "site_id",
     "abo",
     "rh",
     "units"
    ],
    "title": "AppealIn"
   },
...
   "Component": {
    "type": "string",
    "enum": [
     "red_cells",
     "platelets",
     "plasma"
    ],
    "title": "Component"
   },
...
   "DecisionIn": {
    "properties": {
     "approve": {
      "type": "boolean",
      "title": "Approve"
     },
     "units": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Units"
     }
    },
    "type": "object",
    "required": [
     "approve"
    ],
    "title": "DecisionIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MatchOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "donor_ref": {
      "type": "string",
      "title": "Donor Ref"
     },
     "abo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Abo"
       },
       {
        "type": "null"
       }
      ]
     },
     "rh": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Rh"
       },
       {
        "type": "null"
       }
      ]
     },
     "group_verified": {
      "type": "boolean",
      "title": "Group Verified"
     },
     "wave": {
      "type": "integer",
      "title": "Wave"
     },
     "score": {
      "type": "number",
      "title": "Score"
     },
     "status": {
      "$ref": "#/components/schemas/MatchStatus"
     },
     "invited_at": {
      "type": "string",
      "format": "date-time",
      "title": "Invited At"
     },
     "responded_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Responded At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "donor_ref",
     "abo",
     "rh",
     "group_verified",
     "wave",
     "score",
     "status",
     "invited_at",
     "responded_at"
    ],
    "title": "MatchOut"
   },
...
   "PlanOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "horizon_days": {
      "type": "integer",
      "title": "Horizon Days"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "solver_status": {
      "type": "string",
      "title": "Solver Status"
     },
     "solver_seconds": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Solver Seconds"
     },
     "is_fallback": {
      "type": "boolean",
      "title": "Is Fallback"
     },
     "projected": {
      "additionalProperties": true,
      "type": "object",
      "title": "Projected"
     },
     "recommendations": {
      "items": {
       "$ref": "#/components/schemas/RecommendationOut"
      },
      "type": "array",
      "title": "Recommendations"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "horizon_days",
     "status",
     "solver_status",
     "solver_seconds",
     "is_fallback",
     "projected",
     "recommendations"
    ],
    "title": "PlanOut"
   },
...
   "RecommendationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "plan_id": {
      "type": "string",
      "format": "uuid",
      "title": "Plan Id"
     },
     "from_site_id": {
      "type": "string",
      "format": "uuid",
      "title": "From Site Id"
     },
     "to_site_id": {
      "type": "string",
      "format": "uuid",
      "title": "To Site Id"
     },
     "component": {
      "$ref": "#/components/schemas/Component"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "units": {
      "type": "integer",
      "title": "Units"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     },
     "expected_benefit": {
      "additionalProperties": true,
      "type": "object",
      "title": "Expected Benefit"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "decided_by": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided By"
     },
     "decided_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "plan_id",
     "from_site_id",
     "to_site_id",
     "component",
     "abo",
     "rh",
     "units",
     "reason",
     "expected_benefit",
     "status",
     "decided_by",
     "decided_at"
    ],
    "title": "RecommendationOut"
   },
   "RequestIn": {
    "properties": {
     "site_id": {
      "type": "string",
      "format": "uuid",
      "title": "Site Id"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "component": {
      "$ref": "#/components/schemas/Component"
     },
     "units_needed": {
      "type": "integer",
      "maximum": 20.0,
      "minimum": 1.0,
      "title": "Units Needed"
     },
     "urgency": {
      "$ref": "#/components/schemas/Urgency"
     },
     "needed_by": {
      "type": "string",
      "format": "date-time",
      "title": "Needed By"
     }
    },
    "type": "object",
    "required": [
     "site_id",
     "abo",
     "rh",
     "component",
     "units_needed",
     "urgency",
     "needed_by"
    ],
    "title": "RequestIn",
    "description": "FR-REQ-01: hospital and needed-by are required (422 without them)."
   },
   "RequestOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "ref": {
      "type": "string",
      "title": "Ref"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "requester_id": {
      "type": "string",
      "format": "uuid",
      "title": "Requester Id"
     },
     "site_id": {
      "type": "string",
      "format": "uuid",
      "title": "Site Id"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "component": {
      "$ref": "#/components/schemas/Component"
     },
     "units_needed": {
      "type": "integer",
      "title": "Units Needed"
     },
     "units_secured": {
      "type": "integer",
      "title": "Units Secured"
     },
     "urgency": {
      "$ref": "#/components/schemas/Urgency"
     },
     "needed_by": {
      "type": "string",
      "format": "date-time",
      "title": "Needed By"
     },
     "status": {
      "$ref": "#/components/schemas/RequestStatus"
     },
     "confirmed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Confirmed At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "shortfall": {
      "type": "integer",
      "title": "Shortfall"
     },
     "wave": {
      "type": "integer",
      "title": "Wave"
     },
     "escalated_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Escalated At"
     },
     "transfer_plan_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Transfer Plan Id"
     },
     "progress": {
      "$ref": "#/components/schemas/Progress"
     },
     "accepted_matches": {
      "items": {
       "$ref": "#/components/schemas/AcceptedMatch"
      },
      "type": "array",
      "title": "Accepted Matches"
     }
    },
    "type": "object",
    "required": [
     "id",
     "ref",
     "kind",
     "requester_id",
     "site_id",
     "abo",
     "rh",
     "component",
     "units_needed",
     "units_secured",
     "urgency",
     "needed_by",
     "status",
     "confirmed_at",
     "created_at",
     "shortfall",
     "wave",
     "escalated_at",
     "transfer_plan_id",
     "progress",
     "accepted_matches"
    ],
    "title": "RequestOut"
   },
...
   "TransferOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "recommendation_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommendation Id"
     },
     "from_site_id": {
      "type": "string",
      "format": "uuid",
      "title": "From Site Id"
     },
     "to_site_id": {
      "type": "string",
      "format": "uuid",
      "title": "To Site Id"
     },
     "component": {
      "$ref": "#/components/schemas/Component"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "units": {
      "type": "integer",
      "title": "Units"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "dispatched_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dispatched At"
     },
     "received_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Received At"
     },
     "unit_codes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Unit Codes",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "recommendation_id",
     "from_site_id",
     "to_site_id",
     "component",
     "abo",
     "rh",
     "units",
     "status",
     "created_at",
     "dispatched_at",
     "received_at"
    ],
    "title": "TransferOut"
   },
...
   "UnitOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "unit_code": {
      "type": "string",
      "title": "Unit Code"
     },
     "abo": {
      "$ref": "#/components/schemas/Abo"
     },
     "rh": {
      "$ref": "#/components/schemas/Rh"
     },
     "component": {
      "$ref": "#/components/schemas/Component"
     },
     "collected_at": {
      "type": "string",
      "format": "date-time",
      "title": "Collected At"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     },
     "site_id": {
      "type": "string",
      "format": "uuid",
      "title": "Site Id"
     },
     "status": {
      "$ref": "#/components/schemas/UnitStatus"
     },
     "reserved_for": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reserved For"
     }
    },
    "type": "object",
    "required": [
     "id",
     "unit_code",
     "abo",
     "rh",
     "component",
     "collected_at",
     "expires_at",
     "site_id",
     "status",
     "reserved_for"
    ],
    "title": "UnitOut"
   },
...
   "UnitsIn": {
    "properties": {
     "units": {
      "type": "integer",
      "maximum": 20.0,
      "minimum": 1.0,
      "title": "Units"
     }
    },
    "type": "object",
    "required": [
     "units"
    ],
    "title": "UnitsIn"
   },
...
  }
 }
}